clean up.

+++ CoreLib/inc/hnrt/ErrorMessage.h
@@ -1,10 +1,13 @@
 #pragma once
 
 
+#include <Windows.h>
+
+
 namespace hnrt
 {
     struct ErrorMessage
     {
-        static PCWSTR Get(unsigned long code);
+        static PCWSTR Get(DWORD code);
     };
 }
